src: add ns_addrinfo request
As this request is not linked to a specific handle as are the others
that are supported by the library some refactoring has been needed.
A `ns_base_req` has been extracted from `ns_req` and is now the base
class for every request. `ns_req` will remain as the base class for
every linked-to-a-handle request.

PR-URL: https://github.com/nodesource/nsuv/pull/2
Reviewed-by: Trevor Norris <trev.norris@gmail.com>