src: downcast uv handle field from req
Since libuv automatically stores a pointer to the uv_handle_t when the
req is used, downcast the handle field to the ns_handle type. Instead of
storing the handle_ on our own.

It's safer this way since the uv req's handle field is public, and there
are tests that manually override the handle field.

Unfortunately this relies on the user never setting the handle field to
a non ns_handle type. Otherwise it will break everything.

Also add a method to allow overriding the handle field that requires an
ns_handle. Not sure why this should be used, it at least helps prevent
accidentally setting the handle to a non ns_handle.

PR-URL: https://github.com/nodesource/nsuv/pull/16
Reviewed-by: Santiago Gimeno <santiago.gimeno@gmail.com>