src: use macro to check for nullptr cb
Because users can do strange things like cast a nullptr to a void* and
pass that in as a callback, we need to check every callback passed and
make sure to pass nullptr instead of the proxy function as the callback
to the libuv call in case they do.

Make things easier by placing the check in a macro instead of writing it
out by hand every time.

PR-URL: https://github.com/nodesource/nsuv/pull/20