use custom vector allocator to prevent exceptions
Since some programs will compile with -fno-exception and we want zero
chance of this library causing the program to abort, switch to a custom
Allocator for the vector storage of the uv_buf_t.

In order to do this we need to change the API to return the pointer
instead of a reference to the vector.