src: add support for weak_ptr
Allow a weak_ptr to be passed to each callback along with a normal
pointer. This will make it easier to not accidentally hold on to
resources when juggling multiple threads.

Many tests were duplicated into test/*-wp.cc just to make sure the
weak_ptr API has test coverage.

PR-URL: https://github.com/nodesource/nsuv/pull/31
Co-authored-by: Santiago Gimeno <santiago.gimeno@gmail.com>
Reviewed-by: Santiago Gimeno <santiago.gimeno@gmail.com>